+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [0.0.1+7]
+
+* Change the namespaces type from `Map<String, Map<String, Object>> -> Map<String, Object>`
+* Adds I18Next.of(BuildContext) from Localizations
+* Adds `I18NextLocalizationDelegate`
+* Adds convenience methods to `ResourceStore` for adding, removing, and verifiying locales and namespaces
+* Adds asset bundle data source and the LocalizationDataSource interface
+* Changes links to nubank/i18next
+* Adds example app
+
 ## [0.0.1+6]
 
 * Migrated repository to `williamhjcho/i18next`